# app/src/main/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TP1 - Teléfono descompuesto",
    "description": "Title",
    "version": "1.0.0"
  },
  "paths": {
    "/register-node": {
      "post": {
        "operationId": "registerNode",
        "description": "Registra un nuevo nodo",
        "parameters": [
          {
            "name": "host",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "port",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "uuid",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "salt",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Todo bien",
            "headers": {
              "X-Game-Timestamp": {
                "$ref": "#/components/headers/X-Game-Timestamp"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/registerResponse"
                }
              }
            }
          }
        }
      }
    },
    "/unregister-node": {
      "post": {
        "operationId": "unregisterNode",
        "description": "Un participante decide dejar el juego",
        "parameters": [
          {
            "name": "uuid",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "salt",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sashay Away",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/reconfigure": {
      "post": {
        "operationId": "reconfigure",
        "description": "Cambio de siguiente participante",
        "parameters": [
          {
            "name": "uuid",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "salt",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "nextHost",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "nextPort",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "X-Game-Timestamp",
            "in": "header",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cambio aceptado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/relay": {
      "post": {
        "operationId": "relayMessage",
        "description": "Firma un mensaje y lo manda al siguiente",
        "x-content-type": "multipart/form-data",
        "parameters": [
          {
            "name": "X-Game-Timestamp",
            "in": "header",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "El mensaje y las firmas anteriores",
          "content": {
            "multipart/form-data": {
              "encoding": {
                "message": {
                  "contentType": "*/*"
                },
                "signatures": {
                  "contentType": "application/json"
                }
              },
              "schema": {
                "type": "object",
                "required": [
                  "message",
                  "signatures"
                ],
                "properties": {
                  "message": {
                    "$ref": "#/components/schemas/base64String"
                  },
                  "signatures": {
                    "$ref": "#/components/schemas/signatures"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "mensaje recibido y reenviado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/signature"
                }
              }
            }
          },
          "400": {
            "description": "Tu reloj atrasa...",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/signature"
                }
              }
            }
          },
          "503": {
            "description": "Los bits se perdieron en el multiverso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/signature"
                }
              }
            }
          }
        }
      }
    },
    "/play": {
      "post": {
        "operationId": "sendMessage",
        "description": "Comienza el juego!",
        "requestBody": {
          "description": "El mensaje a enviar por la red telefónica",
          "required": true,
          "content": {
            "*/*": {
              "schema": {
                "type": "string",
                "format": "base64"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "La red telefónica funcionó bien!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/playResponse"
                }
              }
            }
          },
          "400": {
            "description": "Juego cerrado. Se llevaron la pelota.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/playResponse"
                }
              }
            }
          },
          "500": {
            "description": "Faltan firmas.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/playResponse"
                }
              }
            }
          },
          "503": {
            "description": "Yo mandé peras y recibí bananas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/playResponse"
                }
              }
            }
          },
          "504": {
            "description": "La red telefónica falló, no me contestaron.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/playResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "headers": {
      "X-Game-Timestamp": {
        "required": true,
        "schema": {
          "type": "integer",
          "minimum": 0
        }
      }
    },
    "schemas": {
      "base64String": {
        "type": "string",
        "format": "base64"
      },
      "registerRequest": {
        "type": "object",
        "required": [
          "host",
          "port",
          "name"
        ],
        "properties": {
          "host": {
            "type": "string"
          },
          "port": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "registerResponse": {
        "type": "object",
        "required": [
          "nextHost",
          "nextPort",
          "timeout"
        ],
        "properties": {
          "nextHost": {
            "type": "string"
          },
          "nextPort": {
            "type": "integer"
          },
          "timeout": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "signature": {
        "type": "object",
        "required": [
          "name",
          "hash"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "El nombre del nodo que firmo"
          },
          "contentType": {
            "type": "string",
            "description": "El hash de la firma del nodo"
          },
          "contentLength": {
            "type": "integer",
            "description": "La longitud del contenido"
          },
          "hash": {
            "type": "string",
            "description": "El hash del contenido calculado por el nodo"
          }
        }
      },
      "signatures": {
        "type": "object",
        "required": ["items"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/signature"
            }
          }
        }
      },
      "playResponse": {
        "type": "object",
        "required": [
          "contentResult",
          "originalContentType",
          "originalLength",
          "originalHash",
          "receivedContentType",
          "receivedLength",
          "receivedHash",
          "signatures"
        ],
        "properties": {
          "contentResult": {
            "type": "string",
            "description": "el resultado del contenido"
          },
          "originalContentType": {
            "type": "string",
            "description": "el tipo de contenido que enviaste"
          },
          "originalLength": {
            "type": "integer",
            "description": "la longitud original que enviaste"
          },
          "originalHash": {
            "type": "string",
            "description": "el hash del mensaje original que enviaste"
          },
          "receivedContentType": {
            "type": "string",
            "description": "el tipo de contenido que volvió"
          },
          "receivedLength": {
            "type": "integer",
            "description": "la longitud que volvió"
          },
          "receivedHash": {
            "type": "string",
            "description": "el hash del mensaje que volvió"
          },
          "signatures": {
            "$ref": "#/components/schemas/signatures"
          }
        }
      }
    }
  }
}